data type change

+++ ProductCrudApplication/src/main/java/com/ProductsProject/ProductCrudApplication/model/Product.java
@@ -16,5 +16,5 @@ public class Product {
 	private String name;
 	private String description;
 	private double price;
-	private int quantityAvailable;
+	private Long quantityAvailable;
 }
